+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## CHANGELOG.md
+## 1.1.0
+* Add absence service
+* Add members service
 
 ## 1.0.0
 * Create React App